started implementing communication between the vst and openutau

+++ VST/src/OpenUTAUPlugin.cpp
@@ -3,8 +3,8 @@
 #include <memory>
 #include <iostream>
 
+
 using namespace OpenUTAU;
-TinyProcessLib::Process *OpenUTAUPlugin::uiProcess = nullptr;
 
 void OpenUTAUPlugin::run(const float **, float **outputs, uint32_t frames, const MidiEvent *midiEvents,
                          uint32_t midiEventCount) {
@@ -14,26 +14,23 @@ void OpenUTAUPlugin::run(const float **, float **outputs, uint32_t frames, const
         std::cout << position.frame << std::endl;
     }
 
-    if (uiProcess) return;
-    std::string openUtauPath = R"(C:\Users\SeleDreams\Dropbox\Documents\Development\OpenUtau\OpenUtau\bin\Debug\netcoreapp3.1)";
-    std::string path = openUtauPath + "/OpenUtau.exe -vst";
-    uiProcess = new TinyProcessLib::Process(path,openUtauPath,[](const char *bytes,size_t n){
-        std::cout << "OpenUtau closed !" << std::endl;
-    });
-
+    if (!uiProcess && !server.isRunning())
+    {
+        /*uiProcess = std::make_unique<TinyProcessLib::Process>(openUtauPath + "/OpenUtau.exe -vst -port 1556",openUtauPath,[](const char *bytes, size_t n){
+            std::cout << "OpenUtau closed !" << std::endl;
+        });*/
+        server.Run();
+        std::cout << "OpenUTAU created! " << openUtauPath << std::endl;
+    }
 }
 
-void OpenUTAUPlugin::initAudioPort(bool input, uint32_t index, AudioPort &port) {
-    Plugin::initAudioPort(input, index, port);
-}
-
-OpenUTAUPlugin::OpenUTAUPlugin() : Plugin(0,0,0){
+OpenUTAUPlugin::OpenUTAUPlugin() : Plugin(0,0,0), server(IPV::V4,1556), uiProcess(nullptr) {
 }
 
 OpenUTAUPlugin::~OpenUTAUPlugin() {
-    if (uiProcess)
+    int exit_status;
+    if (uiProcess && !uiProcess->try_get_exit_status(exit_status))
     {
         uiProcess->kill(true);
-        uiProcess = nullptr;
     }
 }
